@@ -12,47 +12,47 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.4.3",
-    "@angular/cdk": "2.0.0-beta.10",
-    "@angular/common": "^4.2.4",
-    "@angular/compiler": "^4.2.4",
-    "@angular/core": "^4.2.4",
-    "@angular/forms": "^4.2.4",
-    "@angular/http": "^4.2.4",
-    "@angular/material": "2.0.0-beta.10",
-    "@angular/platform-browser": "^4.2.4",
-    "@angular/platform-browser-dynamic": "^4.2.4",
-    "@angular/router": "^4.2.4",
-    "@swimlane/ngx-charts": "^6.0.2",
+    "@angular/animations": "^5.1.3",
+    "@angular/cdk": "5.0.3",
+    "@angular/common": "^5.1.3",
+    "@angular/compiler": "^5.1.3",
+    "@angular/core": "^5.1.3",
+    "@angular/forms": "^5.1.3",
+    "@angular/http": "^5.1.3",
+    "@angular/material": "5.0.3",
+    "@angular/platform-browser": "^5.1.3",
+    "@angular/platform-browser-dynamic": "^5.1.3",
+    "@angular/router": "^5.1.3",
+    "@swimlane/ngx-charts": "^7.0.1",
     "checkdigit": "^1.1.1",
-    "core-js": "^2.4.1",
-    "d3": "^4.11.0",
+    "core-js": "^2.5.3",
+    "d3": "^4.12.2",
     "file-saver": "^1.3.3",
     "hammerjs": "^2.0.8",
-    "rxjs": "^5.4.2",
-    "xlsx": "^0.11.5",
-    "zone.js": "^0.8.14"
+    "rxjs": "^5.5.6",
+    "xlsx": "^0.11.16",
+    "zone.js": "^0.8.19"
   },
   "devDependencies": {
-    "@angular/cli": "1.4.2",
-    "@angular/compiler-cli": "^4.2.4",
-    "@angular/language-service": "^4.2.4",
-    "@types/file-saver": "0.0.1",
-    "@types/jasmine": "~2.5.53",
+    "@angular/cli": "1.6.3",
+    "@angular/compiler-cli": "^5.1.3",
+    "@angular/language-service": "^5.1.3",
+    "@types/file-saver": "1.3.0",
+    "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.1.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "@types/node": "~8.5.7",
+    "codelyzer": "~4.0.2",
+    "jasmine-core": "~2.8.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.0",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "^1.3.3",
+    "karma-jasmine": "^1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "protractor": "~5.2.2",
+    "ts-node": "~4.1.0",
+    "tslint": "~5.8.0",
+    "typescript": "~2.6.2"
   }
 }
