@@ -1,15 +1,17 @@
 import {Injectable} from '@angular/core';
-import {Status} from './status';
 import {HttpClient} from '@angular/common/http';
-import {Promise, defer} from 'q';
+import {defer, Promise} from 'q';
+import Status from '../../common/Status';
+import StatusRequest from '../../common/request/admin/StatusRequest';
 
 @Injectable()
 export class AdminService {
 
   adminPassword: string;
-  result: Status;
+  status: Status;
 
-  constructor(private http: HttpClient) {}
+  constructor(private http: HttpClient) {
+  }
 
   resetAuth(): void {
     this.adminPassword = undefined;
@@ -19,9 +21,9 @@ export class AdminService {
     const deferred = defer<void>();
     this.http.post('http://localhost:3000/api/admin/status', {
       adminPassword: this.adminPassword
-    })
+    } as StatusRequest)
       .subscribe(data => {
-        this.result = <Status>data;
+        this.status = <Status>data;
         deferred.resolve();
       });
 
