@@ -5,7 +5,8 @@ import {ChoiceService} from './choice.service';
 @Injectable()
 export class StudentGuardService implements CanActivate {
 
-  constructor(private choiceService: ChoiceService, private router: Router) {}
+  constructor(private choiceService: ChoiceService, private router: Router) {
+  }
 
   canActivate() {
     if (!this.choiceService.key || !this.choiceService.grade) {
